requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cel-python==0.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.13
h11==0.16.0
idna==3.10
Jinja2==3.1.6
json-logic-qubit==0.9.1
lark==0.12.0
MarkupSafe==3.0.2
pendulum==3.1.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
types-PyYAML==6.0.12.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
